+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Drop support for Node.js < 0.10
+
 2.3.3 / 2022-03-06
 ==================